pyproject.toml
[tool.poetry]
name = "fastapi-kafka-producer"
version = "0.1.0"
description = ""
authors = ["Henri Kestiö <henri.kestio@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
aiokafka = "^0.7.2"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
